Converted client player position and rotation packet
